# PROYECTO/backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentación de la API de Hastechno",
        "description": "API para la gestión de productos, pedidos, usuarios, carritos y más para la tienda virtual Hastechno.",
        "contact": {
            "email": "babeltranp@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/",
            "description": "Servidor de desarrollo local de la API"
        }
    ],
    "paths": {
...
        "/api/facturas/pedido/{idPedido}": {
            "get": {
                "tags": [
                    "Facturas"
                ],
                "summary": "Mostrar factura por ID de pedido",
                "operationId": "41215059f0286380813b236a3b439307",
                "parameters": [
                    {
                        "name": "idPedido",
                        "in": "path",
                        "description": "ID del pedido para el cual se desea obtener la factura.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Factura obtenida correctamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Factura"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Factura no encontrada para el ID de pedido proporcionado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": null,
                                            "nullable": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Factura no encontrada para el pedido."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Cliente": {
                "title": "Cliente",
                "description": "Representa la información de un cliente en el sistema.",
                "properties": {
                    "idCliente": {
                        "description": "ID único del cliente.",
                        "type": "integer",
                        "format": "int64",
                        "readOnly": true,
                        "example": 1
                    },
                    "nombreCliente": {
                        "description": "Nombre del cliente.",
                        "type": "string",
                        "example": "Juan"
                    },
                    "apellidoCliente": {
                        "description": "Apellido del cliente.",
                        "type": "string",
                        "example": "Pérez"
                    },
                    "cedulaCliente": {
                        "description": "Cédula de identidad del cliente.",
                        "type": "string",
                        "example": "1234567890"
                    },
                    "telefonoCliente": {
                        "description": "Número de teléfono del cliente.",
                        "type": "string",
                        "example": "3001234567"
                    },
                    "direccion": {
                        "description": "Dirección de residencia del cliente.",
                        "type": "string",
                        "example": "Calle 10 # 20-30, Bogotá"
                    },
                    "idUsuario": {
                        "description": "ID del usuario asociado a este cliente.",
                        "type": "integer",
                        "format": "int64",
                        "example": 5
                    },
                    "created_at": {
                        "description": "Fecha y hora de creación del registro.",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "description": "Fecha y hora de la última actualización del registro.",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "usuario": {
                        "$ref": "#/components/schemas/Usuario"
                    }
                },
                "type": "object"
            },
            "Factura": {
                "title": "Factura",
                "description": "Representa una factura generada para un pedido, incluyendo el método de pago y la fecha de emisión.",
                "properties": {
                    "idFactura": {
                        "description": "ID único de la factura.",
                        "type": "integer",
                        "format": "int64",
                        "readOnly": true,
                        "example": 1
                    },
                    "idPedido": {
                        "description": "ID del pedido asociado a esta factura.",
                        "type": "integer",
                        "format": "int64",
                        "example": 101
                    },
                    "metodoPago": {
                        "description": "Método de pago utilizado para el pedido.",
                        "type": "string",
                        "example": "Tarjeta de Crédito"
                    },
                    "fechaFactura": {
                        "description": "Fecha de emisión de la factura (formato YYYY-MM-DD).",
                        "type": "string",
                        "format": "date",
                        "example": "2024-07-07"
                    },
                    "created_at": {
                        "description": "Fecha y hora de creación de la factura.",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "description": "Fecha y hora de la última actualización de la factura.",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "pedido": {
                        "$ref": "#/components/schemas/Pedido"
                    }
                },
                "type": "object"
            },
            "Marca": {
                "title": "Marca",
                "description": "Representa la marca de un producto.",
                "properties": {
                    "idMarca": {
                        "description": "ID único de la marca.",
                        "type": "integer",
                        "format": "int64",
                        "readOnly": true,
                        "example": 1
                    },
                    "nombreMarca": {
                        "description": "Nombre de la marca.",
                        "type": "string",
                        "example": "Samsung"
                    },
                    "created_at": {
                        "description": "Fecha y hora de creación de la marca.",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "description": "Fecha y hora de la última actualización de la marca.",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            },
            "Pedido": {
                "title": "Pedido",
                "description": "Representa un pedido realizado por un cliente, incluyendo la fecha y los productos asociados.",
                "properties": {
                    "idPedido": {
                        "description": "ID único del pedido.",
                        "type": "integer",
                        "format": "int64",
                        "readOnly": true,
                        "example": 1
                    },
                    "idCliente": {
                        "description": "ID del cliente que realizó el pedido.",
                        "type": "integer",
                        "format": "int64",
                        "example": 10
                    },
                    "fechaPedido": {
                        "description": "Fecha en que se realizó el pedido (formato YYYY-MM-DD).",
                        "type": "string",
                        "format": "date",
                        "example": "2024-07-07"
                    },
                    "created_at": {
                        "description": "Fecha y hora de creación del pedido.",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "description": "Fecha y hora de la última actualización del pedido.",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "cliente": {
                        "$ref": "#/components/schemas/Cliente"
                    },
                    "productos": {
                        "description": "Lista de productos incluidos en el pedido (cargado con eager loading).",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PedidoProducto"
                        }
                    },
                    "factura": {
                        "$ref": "#/components/schemas/Factura"
                    }
                },
                "type": "object"
            },
            "PedidoProducto": {
                "title": "Pedido Producto",
                "description": "Representa un item de producto dentro de un pedido, detallando la cantidad y el valor total de ese producto en particular.",
                "properties": {
                    "id": {
                        "description": "ID único del registro de PedidoProducto (clave primaria de la tabla pivote).",
                        "type": "integer",
                        "format": "int64",
                        "readOnly": true,
                        "example": 1
                    },
                    "idPedido": {
                        "description": "ID del pedido al que pertenece este item.",
                        "type": "integer",
                        "format": "int64",
                        "example": 101
                    },
                    "idProducto": {
                        "description": "ID del producto asociado a este item.",
                        "type": "integer",
                        "format": "int64",
                        "example": 205
                    },
                    "cantidadProducto": {
                        "description": "Cantidad de este producto en el pedido.",
                        "type": "integer",
                        "example": 2
                    },
                    "valorTotal": {
                        "description": "Valor total de este producto (cantidad * precio unitario) dentro del pedido.",
                        "type": "number",
                        "format": "float",
                        "example": 2400
                    },
                    "created_at": {
                        "description": "Fecha y hora de creación del registro.",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "description": "Fecha y hora de la última actualización del registro.",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "pedido": {
                        "$ref": "#/components/schemas/Pedido"
                    },
                    "producto": {
                        "$ref": "#/components/schemas/Producto"
                    }
                },
                "type": "object"
            },
            "Producto": {
                "title": "Producto",
                "description": "Representa un producto disponible en el inventario de la tienda.",
                "properties": {
                    "idProducto": {
                        "description": "ID único del producto.",
                        "type": "integer",
                        "format": "int64",
                        "readOnly": true,
                        "example": 1
                    },
                    "nombreProducto": {
                        "description": "Nombre del producto.",
                        "type": "string",
                        "example": "Laptop Ultrabook X1"
                    },
                    "definicion": {
                        "description": "Descripción detallada del producto.",
                        "type": "string",
                        "example": "Laptop de alto rendimiento con procesador i7, 16GB RAM y 512GB SSD."
                    },
                    "valorProducto": {
                        "description": "Precio del producto.",
                        "type": "number",
                        "format": "float",
                        "example": 1200.5
                    },
                    "idProveedor": {
                        "description": "ID del proveedor del producto.",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "disponibilidad": {
                        "description": "Indica si el producto está disponible para la venta.",
                        "type": "boolean",
                        "example": true
                    },
                    "cantidadStock": {
                        "description": "Cantidad de unidades del producto en stock.",
                        "type": "integer",
                        "example": 50
                    },
                    "idMarca": {
                        "description": "ID de la marca del producto.",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "created_at": {
                        "description": "Fecha y hora de creación del producto.",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "description": "Fecha y hora de la última actualización del producto.",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "marca": {
                        "$ref": "#/components/schemas/Marca"
                    },
                    "proveedor": {
                        "$ref": "#/components/schemas/ProveedorBasic"
                    }
                },
                "type": "object"
            },
            "ProveedorBasic": {
                "title": "Proveedor Básico",
                "description": "Información básica de un proveedor, usada en el contexto de un producto.",
                "properties": {
                    "idProveedor": {
                        "description": "ID único del proveedor.",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "nombreProveedor": {
                        "description": "Nombre del proveedor.",
                        "type": "string",
                        "example": "Tech Supplies Inc."
                    }
                },
                "type": "object"
            },
...
            "Rol": {
                "title": "Rol",
                "description": "Representa el rol de un usuario en el sistema (ej. SuperAdmin, Admin, Cliente).",
                "properties": {
                    "idRol": {
                        "description": "ID único del rol.",
                        "type": "integer",
                        "format": "int64",
                        "readOnly": true,
                        "example": 1
                    },
                    "nombreRol": {
                        "description": "Nombre del rol.",
                        "type": "string",
                        "example": "SuperAdmin"
                    },
                    "created_at": {
                        "description": "Fecha y hora de creación del rol.",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "description": "Fecha y hora de la última actualización del rol.",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            },
            "Usuario": {
                "title": "Usuario",
                "description": "Representa un usuario autenticable en el sistema, incluyendo su rol y relaciones con perfiles específicos (Admin, Cliente).",
                "properties": {
                    "idUsuario": {
                        "description": "ID único del usuario.",
                        "type": "integer",
                        "format": "int64",
                        "readOnly": true,
                        "example": 1
                    },
                    "idRol": {
                        "description": "ID del rol asociado al usuario (ej. 1 para SuperAdmin, 2 para Admin, 3 para Cliente).",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "email": {
                        "description": "Correo electrónico único del usuario.",
                        "type": "string",
                        "format": "email",
                        "example": "usuario@example.com"
                    },
                    "password": {
                        "description": "Contraseña del usuario (solo para entrada, no se expone en la salida).",
                        "type": "string",
                        "format": "password",
                        "writeOnly": true,
                        "example": "password123"
                    },
                    "created_at": {
                        "description": "Fecha y hora de creación del usuario.",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "description": "Fecha y hora de la última actualización del usuario.",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "rol": {
                        "$ref": "#/components/schemas/Rol"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingresa tu token JWT con el prefijo 'Bearer ' (ej: 'Bearer eyJ0eXAiOiJKV1Q...') para acceder a los endpoints protegidos.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}